backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# AI / ML
easyocr==1.7.1
ollama==0.2.1
openai==1.30.1
sentence-transformers==3.0.1
langchain==0.2.5
langchain-community==0.2.5
langchain-huggingface==0.0.3
chromadb==0.5.0

# Image processing
Pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
pydantic==2.7.1
pydantic-settings==2.3.0

# OCR fallback
pytesseract==0.3.10
